requirements.txt
bs4
beautifulsoup4
requests==2.3.0
Pillow==2.5.3
GroupyAPI == 0.7.1
Flask==0.12
gunicorn==19.6.0

